Made that change+
Updated the expected_hash for the teams_data and made sure the test read 'teams_data" not "game_data" again--duplication otherwise. Also changed the expected_hash for game_data. We need to go through and change these for all the other tests/change where we use merge_hashes_module, but this is on the right track. I say let's merge now and we'll deal with the conflicts later--easier to edit in atom than this awful thing. haha